requirements.txt
apscheduler>=3.6,<4.0
gunicorn>=20.0,<21.0
hug>=2.6.0,<3.0
prometheus_client>=0.7.1,<1.0
pyyaml>=5.1,<6.0
requests>=2.19,<3.0
